+++ contracts/mocks/metatx/ERC20MetaMock.sol
@@ -1,4 +1,4 @@
-pragma solidity = 0.6.2;
+pragma solidity ^0.6.6;
 
 import "../../token/ERC20/ERC20Base.sol";
 import "../../metatx/ERC20Fees.sol";